Continued development...

+++ admin/includes/view_all_posts.php
@@ -90,7 +90,7 @@
 						echo "<td>{$post_tags}</td>";
 						echo "<td>{$post_comment_count}</td>";
 						echo "<td>{$post_date}</td>";
-						echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}' class='btn btn-warning' role='button'>Edit</a>&nbsp;&nbsp;<a href='posts.php?delete={$post_id}' class='btn btn-danger' role='button'>Delete</a></td>";
+						echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}' class='btn btn-warning' role='button'>Edit</a>&nbsp;&nbsp;<a onClick=\"javascript: return confirm('Are you sure?');\" href='posts.php?delete={$post_id}' class='btn btn-danger' role='button'>Delete</a></td>";
 						echo "</tr>";
 					}
